Successful code refactor
